Detail Bayslot logic where service time exceeds more than a day. Sales Process...

See merge request telliant-mammoth/strive!485

+++ Sourcecode/API/Strive.BusinessEntities/Model/EmployeeHourlyRate.cs
@@ -13,14 +13,14 @@ namespace Strive.BusinessEntities.Model
         [Column]
         public int? EmployeeId { get; set; }
 
-       [Column]
+        [Column]
         public int? RoleId { get; set; }
 
         [Column]
         public int? LocationId { get; set; }
 
         [Column]
-        public int HourlyRate{ get; set; }
+        public decimal HourlyRate{ get; set; }
        
         [Column]
         public bool? IsActive { get; set; }
@@ -39,6 +39,6 @@ namespace Strive.BusinessEntities.Model
 
         [Column]
         public DateTimeOffset? UpdatedDate { get; set; }
-    
+
     }
 }
